requirements.txt
annotated-types==0.7.0
Authlib==1.3.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dparse==0.6.4b0
filelock==3.12.4
idna==3.10
jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
packaging==24.1
pdfminer.six==20240706
pillow==10.4.0
psutil==6.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.18.0
pypdf2==3.0.1
reportlab==4.2.2
requests==2.32.3
rich==13.8.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safety==3.2.7
safety-schemas==0.0.5
shellingham==1.5.4
tomli==2.0.1
typer==0.12.5
typing-extensions==4.12.2
urllib3==2.2.3
